--- src/main/python/plotlyst/core/sprint.py
@@ -1,67 +1,0 @@
-"""
-Plotlyst
-Copyright (C) 2021-2024  Zsolt Kovari
-
-This file is part of Plotlyst.
-
-Plotlyst is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Plotlyst is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-"""
-from PyQt6.QtCore import QObject, pyqtSignal, QTimer
-
-
-class TimerModel(QObject):
-    DefaultValue: int = 60 * 5
-
-    valueChanged = pyqtSignal()
-    finished = pyqtSignal()
-
-    def __init__(self, parent=None):
-        super(TimerModel, self).__init__(parent)
-        self.value: int = self.DefaultValue
-
-        self._timer = QTimer()
-        self._timer.setInterval(1000)
-        self._timer.timeout.connect(self._tick)
-
-    def start(self, value: int):
-        if value == 3600:
-            value -= 1
-        self.value = value
-        self._timer.start()
-
-    def stop(self):
-        self._timer.stop()
-        self.value = self.DefaultValue
-
-    def remainingTime(self):
-        minutes = self.value // 60
-        seconds = self.value % 60
-        return minutes, seconds
-
-    def isActive(self) -> bool:
-        return self._timer.isActive()
-
-    def toggle(self):
-        if self._timer.isActive():
-            self._timer.stop()
-        else:
-            self._timer.start()
-
-    def _tick(self):
-        if self.value == 0:
-            self._timer.stop()
-            self.finished.emit()
-        else:
-            self.value -= 1
-            self.valueChanged.emit()
